# Improve performance of copy