--- a/tests/test_copy.py
+++ b/tests/test_copy.py
@@ -643,7 +643,7 @@ def test_worker_life(conn, format, buffer):
 
 def test_worker_error_propagated(conn, monkeypatch):
 
-    def copy_to_broken(pgconn, buffer):
+    def copy_to_broken(pgconn, buffer, flush=True):
         raise ZeroDivisionError
         yield