Fixes #745 

I kept the flush code the same as `copy_end`, but think it may be better to break if `f != 1` instead of `f == 0` in both places to account for errors, which return -1 ([Postgres docs](https://www.postgresql.org/docs/current/libpq-async.html#LIBPQ-PQFLUSH)).