@dvarrazzo 2024-03-04
Errors don't need to be accounted for because `PGconn.flush()` will raise an exception if `PQflush()` returns -1.

https://github.com/psycopg/psycopg/blob/558cbe7ef7ef6a6d1f6c65e368aaaf9f9a10108e/psycopg/psycopg/pq/pq_ctypes.py#L588