@dvarrazzo 2024-03-04
Thank you for your MR. I need to review the whole copy pipeline because there were several considerations we made in the past (around buffering, for instance) that apparently don't hold all the water I thought they did and maybe there is code around that that can be removed.