@dvarrazzo 2024-05-03
Hello,

on Linux, I struggle to see any difference between using or not using flush. But I do accept that on macOS it might be different.

I have added a minimal framework that can be used to test copy performance. Could you please give it a try and report what sort of time difference you have with or without your patch, on your platform? 

Can you also report if you are testing on a network connection or if your database is on localhost?

Feel free to make any change you think useful to the test script. Thank you very much!