@ankane 2024-05-06
Hi @dvarrazzo, I created a [CI workflow](https://github.com/ankane/psycopg-repro2/blob/master/.github/workflows/build.yml) to show the difference (and can be used to test). The database is on localhost.

Mac - [CI logs](https://github.com/ankane/psycopg-repro2/actions/runs/8962145231/job/24610642251)

```sh
tests/scripts/copytest.py --nrecs 1000000
```

- master: `time to copy: 22.962515 sec`
- PR: `time to copy: 1.659219 sec`

Linux - private runner - [CI logs](https://github.com/ankane/psycopg-repro2/actions/runs/8962145231/job/24610642397)

```sh
tests/scripts/copytest.py --colsize 200 --nrecs 1000000
```

- master: `time to copy: 26.573821 sec`
- PR: `time to copy: 17.158544 sec`

Linux - public runner - [CI logs](https://github.com/ankane/psycopg-repro2/actions/runs/8962145231/job/24610774796)

```sh
tests/scripts/copytest.py --colsize 200 --nrecs 1000000
```

- master: `time to copy: 11.504595 sec`
- PR: `time to copy: 12.674765 sec`

For some reason, the PR is faster than master on CI when the repo is private, but slightly slower on CI when the repo is public.