@dvarrazzo 2024-05-06
Thank you @ankane. It looks like the result is pretty clear on macOS, but less so on Linux. You also had results where the performance on linux seemed improved in previous version of you message, right?

For me the right target is 3.1.18 because I would like to release a change in 3.1.19 if it's not excessively invasive.

Maybe we can flush at every row for macOS only. I wonder what we should do for Windows. Is it easy to add it to your workflow?