@ankane 2024-05-06
That seems reasonable. Another approach that might work for all platforms would be to keep track of the buffer size and only flush once it hits a certain threshold (so it's not doing it every row).

Here are the [CI results](https://github.com/ankane/psycopg-repro2/actions/runs/8973183518/job/24642772251) for Windows:

```sh
python tests/scripts/copytest.py --colsize 100 --nrecs 1000000
```

- w/ patch: 17.495542 sec
- w/o patch: 16.414442 sec