@dvarrazzo 2024-05-07
Ok, given these numbers I will merge this MR with flush only activated on macOS. We can tweak it further in the future, but it seems that at the moment we are hitting some quadratic behaviour on macOS and the most important thing is to avoid them.